+++ src/main/java/kr/hhplus/be/commerce/domain/order/OrderItem.java
@@ -31,22 +31,22 @@ public class OrderItem extends AuditingFields implements Serializable {
 
     @Column(name = "quantity", nullable = false)
     @Comment("주문 수량")
-    private Integer quantity;
+    private Long quantity;
 
     @Column(name = "price", nullable = false)
     @Comment("상품 주문 당시 가격")
-    private Integer price;
+    private Long price;
 
     @Column(name = "total_price", nullable = false)
     @Comment("주문 항목 총 금액")
-    private Integer totalPrice;
+    private Long totalPrice;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "order_id", referencedColumnName = "order_id", nullable = false)
     @Comment("주문")
     private Order order;
 
-    public static OrderItem create(Long productId, Integer quantity, Integer price, Integer totalPrice) {
+    public static OrderItem create(Long productId, Long quantity, Long price, Long totalPrice) {
         OrderItem entity = new OrderItem();
         entity.productId = productId;
         entity.quantity = quantity;
