+++ src/main/java/kr/hhplus/be/commerce/interfaces/dto/responseDto/TopProductResponse.java
@@ -6,8 +6,8 @@ import lombok.Builder;
 public record TopProductResponse(
         Long id,
         String name,
-        Integer price,
+        Long price,
         String description,
-        Integer soldQuantity
+        Long soldQuantity
 ) {
 }
